Add profile annotations for dev environment
