+++ application.properties
@@ -23,4 +23,5 @@ security.oauth2.client.tmc.scopes=public
 security.oauth2.client.tmc.client-name=TMC
 security.oauth2.client.tmc.client-alias=tmc
 
-production=false
+# Profiles
+spring.profiles.active=dev
