+++ src/main/java/pingis/config/SecurityConfig.java
@@ -2,8 +2,9 @@ package pingis.config;
 
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
+import org.springframework.core.annotation.Order;
 import org.springframework.http.HttpMethod;
-import org.springframework.security.config.annotation.authentication.builders.AuthenticationManagerBuilder;
 import org.springframework.security.config.annotation.web.builders.HttpSecurity;
 import org.springframework.security.config.annotation.web.builders.WebSecurity;
 import org.springframework.security.config.annotation.web.configuration.EnableWebSecurity;
@@ -12,6 +13,7 @@ import org.springframework.security.oauth2.client.registration.ClientRegistratio
 import org.springframework.security.oauth2.client.registration.ClientRegistrationProperties;
 import org.springframework.security.web.authentication.LoginUrlAuthenticationEntryPoint;
 
+@Profile({"prod", "oauth"})
 @EnableWebSecurity
 public class SecurityConfig extends WebSecurityConfigurerAdapter {
     @Override
@@ -34,9 +36,7 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
                 .authenticationEntryPoint(new LoginUrlAuthenticationEntryPoint("/login"))
             .and()
             .oauth2Login()
-                .clients(tmcClientRegistration())
-            .and()
-            .formLogin().loginPage("/login");
+                .clients(tmcClientRegistration());
     }
 
     //Registers TMC's information for the application
@@ -48,14 +48,5 @@ public class SecurityConfig extends WebSecurityConfigurerAdapter {
 
         tmcClientRegistration = new ClientRegistration.Builder(clientRegistrationProperties).build();
         return new ClientRegistration[]{tmcClientRegistration};
-    }
-    
-    @Autowired
-    public void configureGlobal(AuthenticationManagerBuilder auth) throws Exception {
-        auth
-            .inMemoryAuthentication()
-                .withUser("user").password("password").roles("USER")
-                .and()
-                .withUser("admin").password("password").roles("ADMIN");
     }
 }
