+++ src/main/java/pingis/controllers/ChallengeController.java
@@ -9,6 +9,7 @@ import org.springframework.web.servlet.view.RedirectView;
 import pingis.entities.Task;
 import pingis.entities.Challenge;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.context.annotation.Profile;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -23,6 +24,7 @@ import pingis.utils.JavaSyntaxChecker;
 
 @Controller
 public class ChallengeController {
+
     @Autowired
     private ChallengeRepository cr;
 
@@ -35,6 +37,7 @@ public class ChallengeController {
     public String hello() {
         return "hello";
     }
+
     //CHECKSTYLE:OFF
     @RequestMapping(value = "/task/{challengeId}/{taskId}/{taskType}", method = RequestMethod.GET)
     public String task(Model model, @PathVariable String taskType,
@@ -108,6 +111,5 @@ public class ChallengeController {
         model.addAttribute("feedback", "Good work!");
         return "feedback";
     }
-
 
 }
