+++ src/main/java/pingis/controllers/UserDevController.java
@@ -1,19 +1,18 @@
 package pingis.controllers;
 
 import org.springframework.context.annotation.Profile;
-import org.springframework.core.annotation.Order;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.Model;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 
-@Profile({"prod", "oauth"})
+@Profile("dev")
 @Controller
-public class UserController {
+public class UserDevController {
     
     @RequestMapping(value = "/login", method = RequestMethod.GET)
     public String login(Model model) {
-        return "redirect:/oauth2/authorize/code/tmc";
+        return "login";
     }
 
     @RequestMapping(value = "/user", method = RequestMethod.GET)
